+++ lib/src/iregexp.dart
@@ -7,22 +7,22 @@ class IRegexp {
   /// Creates an instance from a string. The [pattern] is parsed once, and
   /// the instance may be used many times after that.
   IRegexp(this.pattern)
-      : regexp = IRegexpGrammarDefinition.parser.parse(pattern).value;
+      : _regexp = IRegexpGrammarDefinition.parser.parse(pattern).value;
 
   /// The pattern used to create this instance.
   final String pattern;
 
   /// The pattern converted to a RegExp-compatible expression.
-  final String regexp;
+  final String _regexp;
 
   RegExp? _regExpCached, _substringRegExpCached;
 
   /// Returns a [RegExp] which matches the same strings as this [pattern].
-  RegExp toRegExp() => _regExpCached ?? (_regExpCached = _regExp('^$regexp\$'));
+  RegExp toRegExp() => _regExpCached ?? (_regExpCached = _regExp('^$_regexp\$'));
 
   /// Returns a [RegExp] which matches any substrings matched by [pattern].
   RegExp toSubstringRegExp() =>
-      _substringRegExpCached ?? (_substringRegExpCached = _regExp(regexp));
+      _substringRegExpCached ?? (_substringRegExpCached = _regExp(_regexp));
 
   /// Checks whether this [pattern] matches the [input].
   bool matches(String input) => toRegExp().hasMatch(input);